+++ graphql-compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/PluginUtils.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.stdlib.graphql.compiler;
+
+import io.ballerina.compiler.api.SemanticModel;
+import io.ballerina.compiler.api.symbols.MethodSymbol;
+import io.ballerina.compiler.api.symbols.ModuleSymbol;
+import io.ballerina.compiler.api.symbols.Qualifier;
+import io.ballerina.compiler.api.symbols.Symbol;
+import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
+import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
+import io.ballerina.stdlib.graphql.compiler.PluginConstants.CompilationErrors;
+import io.ballerina.tools.diagnostics.Diagnostic;
+import io.ballerina.tools.diagnostics.DiagnosticFactory;
+import io.ballerina.tools.diagnostics.DiagnosticInfo;
+import io.ballerina.tools.diagnostics.Location;
+
+import java.util.Optional;
+
+/**
+ * Util class for the compiler plugin.
+ */
+public class PluginUtils {
+    public static boolean validateModuleId(ModuleSymbol moduleSymbol) {
+        String moduleName = moduleSymbol.id().moduleName();
+        String orgName = moduleSymbol.id().orgName();
+        return moduleName.equals(PluginConstants.PACKAGE_PREFIX) &&
+                orgName.equals(PluginConstants.PACKAGE_ORG);
+    }
+
+    public static MethodSymbol getMethodSymbol(SyntaxNodeAnalysisContext context,
+                                               FunctionDefinitionNode functionDefinitionNode) {
+        MethodSymbol methodSymbol = null;
+        SemanticModel semanticModel = context.semanticModel();
+        Optional<Symbol> symbol = semanticModel.symbol(functionDefinitionNode);
+        if (symbol.isPresent()) {
+            methodSymbol = (MethodSymbol) symbol.get();
+        }
+        return methodSymbol;
+    }
+
+    public static boolean isRemoteFunction(SyntaxNodeAnalysisContext context,
+                                           FunctionDefinitionNode functionDefinitionNode) {
+        MethodSymbol methodSymbol = getMethodSymbol(context, functionDefinitionNode);
+        return methodSymbol.qualifiers().contains(Qualifier.REMOTE);
+    }
+
+    public static void updateContext(SyntaxNodeAnalysisContext context, CompilationErrors errorCode,
+                                     Location location) {
+        DiagnosticInfo diagnosticInfo = new DiagnosticInfo(
+                errorCode.getErrorCode(), errorCode.getError(), errorCode.getDiagnosticSeverity());
+        Diagnostic diagnostic = DiagnosticFactory.createDiagnostic(diagnosticInfo, location);
+        context.reportDiagnostic(diagnostic);
+    }
+}